+++ mapwize-ui/src/main/java/io/mapwize/mapwizeui/MapwizeFragment.java
@@ -15,6 +15,12 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.fragment.app.Fragment;
+
+import com.mapbox.mapboxsdk.Mapbox;
+import com.mapbox.mapboxsdk.maps.MapboxMapOptions;
+
+import java.util.List;
+
 import io.mapwize.mapwizesdk.api.ApiCallback;
 import io.mapwize.mapwizesdk.api.Direction;
 import io.mapwize.mapwizesdk.api.DirectionMode;
@@ -22,8 +28,9 @@ import io.mapwize.mapwizesdk.api.DirectionPoint;
 import io.mapwize.mapwizesdk.api.Place;
 import io.mapwize.mapwizesdk.core.MapwizeConfiguration;
 import io.mapwize.mapwizesdk.map.MapOptions;
+import io.mapwize.mapwizesdk.map.MapwizeIndoorLocation;
 
-public class MapwizeFragment extends Fragment {
+public class MapwizeFragment extends Fragment implements MapwizeFragmentDirectionsInterface {
 
     // Options
     private static String ARG_OPTIONS = "param_options";
@@ -331,4 +338,54 @@ public class MapwizeFragment extends Fragment {
         }
     }
 
+
+    // MapwizeFragmentDirectionsInterface implementation
+
+
+    @Override
+    public void showDirectionLoading() {
+        mapwizeUIView.showDirectionLoading();
+    }
+
+    @Override
+    public void showDirectionInfo(Direction direction) {
+        mapwizeUIView.showDirectionInfo(direction);
+    }
+
+    @Override
+    public void showAccessibleDirectionModes(List<DirectionMode> modes) {
+        mapwizeUIView.showAccessibleDirectionModes(modes);
+    }
+
+    @Override
+    public void showDirectionError() {
+        mapwizeUIView.showDirectionError();
+    }
+
+    @Override
+    public void hideInfo() {
+        mapwizeUIView.hideInfo();
+    }
+
+    @Nullable
+    @Override
+    public MapwizeIndoorLocation getUserLocation() {
+        return mapwizeUIView.getUserLocation();
+    }
+
+    @NonNull
+    @Override
+    public List<DirectionMode> getDirectionModes() {
+        return mapwizeUIView.getDirectionModes();
+    }
+
+    @Override
+    public void setDirection(Direction direction) {
+        mapwizeUIView.setDirection(direction);
+    }
+
+    @Override
+    public void quitDirections() {
+        mapwizeUIView.quitDirections();
+    }
 }