+++ mapwize-ui/src/main/java/io/mapwize/mapwizeui/MapwizeFragmentUISettings.java
@@ -13,12 +13,18 @@ public class MapwizeFragmentUISettings implements Parcelable {
     private boolean followUserButtonHidden;
     private boolean floorControllerHidden;
     private boolean compassHidden;
+    private boolean universesButtonHidden;
+    private boolean directionsQrCodeHidden;
+    private boolean closestExitButtonHidden;
 
-    private MapwizeFragmentUISettings(boolean menuButtonHidden, boolean followUserButtonHidden, boolean floorControllerHidden, boolean compassHidden) {
+    private MapwizeFragmentUISettings(boolean menuButtonHidden, boolean followUserButtonHidden, boolean floorControllerHidden, boolean compassHidden, boolean universesButtonHidden, boolean directionsQrCodeHidden, boolean closestExitButtonHidden) {
         this.menuButtonHidden = menuButtonHidden;
         this.followUserButtonHidden = followUserButtonHidden;
         this.floorControllerHidden = floorControllerHidden;
         this.compassHidden = compassHidden;
+        this.universesButtonHidden = universesButtonHidden;
+        this.directionsQrCodeHidden = directionsQrCodeHidden;
+        this.closestExitButtonHidden = closestExitButtonHidden;
     }
 
     public boolean isMenuButtonHidden() {
@@ -37,11 +43,26 @@ public class MapwizeFragmentUISettings implements Parcelable {
         return compassHidden;
     }
 
+    public boolean isUniversesButtonHidden() {
+        return universesButtonHidden;
+    }
+
+    public boolean isDirectionsQrCodeHidden() {
+        return directionsQrCodeHidden;
+    }
+
+    public boolean isClosestExitButtonHidden() {
+        return closestExitButtonHidden;
+    }
+
     private MapwizeFragmentUISettings(Parcel in) {
         menuButtonHidden = in.readByte() != 0;
         followUserButtonHidden = in.readByte() != 0;
         floorControllerHidden = in.readByte() != 0;
         compassHidden = in.readByte() != 0;
+        universesButtonHidden = in.readByte() != 0;
+        directionsQrCodeHidden = in.readByte() != 0;
+        closestExitButtonHidden = in.readByte() != 0;
     }
 
     public static final Creator<MapwizeFragmentUISettings> CREATOR = new Creator<MapwizeFragmentUISettings>() {
@@ -67,6 +88,9 @@ public class MapwizeFragmentUISettings implements Parcelable {
         dest.writeByte((byte) (followUserButtonHidden ? 1 : 0));
         dest.writeByte((byte) (floorControllerHidden ? 1 : 0));
         dest.writeByte((byte) (compassHidden ? 1 : 0));
+        dest.writeByte((byte) (universesButtonHidden ? 1 : 0));
+        dest.writeByte((byte) (directionsQrCodeHidden ? 1 : 0));
+        dest.writeByte((byte) (closestExitButtonHidden ? 1 : 0));
     }
 
     @Override
@@ -76,6 +100,9 @@ public class MapwizeFragmentUISettings implements Parcelable {
                 ", followUserButtonHidden=" + followUserButtonHidden +
                 ", floorControllerHidden=" + floorControllerHidden +
                 ", compassHidden=" + compassHidden +
+                ", universesButtonHidden=" + universesButtonHidden +
+                ", directionsQrCodeHidden=" + directionsQrCodeHidden +
+                ", closestExitButtonHidden=" + closestExitButtonHidden +
                 '}';
     }
 
@@ -85,12 +112,18 @@ public class MapwizeFragmentUISettings implements Parcelable {
         private boolean followUserButtonHidden;
         private boolean floorControllerHidden;
         private boolean compassHidden;
+        private boolean universesButtonHidden;
+        private boolean directionsQrCodeHidden;
+        private boolean closestExitButtonHidden;
 
         public Builder() {
             this.menuButtonHidden = false;
             this.followUserButtonHidden = false;
             this.floorControllerHidden = false;
             this.compassHidden = false;
+            this.universesButtonHidden = false;
+            this.directionsQrCodeHidden = false;
+            this.closestExitButtonHidden = false;
         }
 
         /**
@@ -134,12 +167,43 @@ public class MapwizeFragmentUISettings implements Parcelable {
         }
 
         /**
+         * Show/Hide the universes button
+         * @param isHidden true if you want to hide the universes button
+         * @return the builder
+         */
+        public Builder universesButtonHidden(boolean isHidden) {
+            this.universesButtonHidden = isHidden;
+            return this;
+        }
+
+        /**
+         * Show/Hide the qr code button in the search bar
+         * @param isHidden true if you want to hide the qr code button in the search bar
+         * @return the builder
+         */
+        public Builder poiQrCodeButtonHidden(boolean isHidden) {
+            this.directionsQrCodeHidden = isHidden;
+            return this;
+        }
+
+        /**
+         * Show/Hide the closest exit button in the UI
+         * @param isHidden true if you want to hide the closest exit button in the UI
+         * @return the builder
+         */
+        public Builder closestExitButtonHidden(boolean isHidden) {
+            this.closestExitButtonHidden = isHidden;
+            return this;
+        }
+
+
+
+        /**
          * Build the fragment UISettings
          * @return the MapwizeFragmentUISettings
          */
         public MapwizeFragmentUISettings build() {
-            return new MapwizeFragmentUISettings(this.menuButtonHidden, this.followUserButtonHidden, this.floorControllerHidden, this.compassHidden);
+            return new MapwizeFragmentUISettings(this.menuButtonHidden, this.followUserButtonHidden, this.floorControllerHidden, this.compassHidden, this.universesButtonHidden, this.directionsQrCodeHidden, this.closestExitButtonHidden);
         }
-
     }
 }