+++ changelog.md
@@ -1,6 +1,25 @@
 # UI Components Changelog
 
 
+## 2.6.0
+
+- Using floor translation in search result and place details view
+- Updating sdk to version `3.7.0`.
+
+## 2.5.3
+
+- Using a floor object instead of a number in `PlaceDetails`.
+- Updating sdk to version `3.6.3`.
+
+
+## 2.5.2
+
+-  Setting occupancy to available if events are null.
+
+## 2.5.1
+
+-  Adding preventExpandDetails option to PlaceDetailsConfig.
+
 ## 2.5.0
 
 - Adding place details UI.